requirements.txt
streamlit==1.28.0
pandas==2.1.4
numpy==1.24.4
plotly==5.15.0
yfinance==0.2.40
pymongo==4.5.0
scipy==1.11.0
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.23.0
asyncio-throttle==1.0.2
certifi==2023.7.22
motor==3.3.0
